Moving choices processing into choices module
