+++ test/choices_test.py
@@ -2,12 +2,14 @@
 
 import pytest
 
+from brewtils.errors import PluginParamError
+
 try:
     from lark import ParseError
 except ImportError:
     from lark.common import ParseError
 
-from brewtils.choices import parse
+from brewtils.choices import parse, process_choices
 
 
 class TestChoices(object):
@@ -130,3 +132,129 @@ class TestChoices(object):
     )
     def test_parse_no_hint(self, input_string, expected):
         assert expected == parse(input_string)
+
+
+class TestProcessChoices(object):
+    @pytest.fixture
+    def cmd(self):
+        class Bar(object):
+            def cmd(self, foo):
+                """Docstring"""
+                return foo
+
+        return Bar.cmd
+
+    @pytest.mark.parametrize(
+        "choices,expected",
+        [
+            (
+                ["1", "2", "3"],
+                {
+                    "type": "static",
+                    "value": ["1", "2", "3"],
+                    "display": "select",
+                    "strict": True,
+                },
+            ),
+            (
+                list(range(100)),
+                {
+                    "type": "static",
+                    "value": list(range(100)),
+                    "display": "typeahead",
+                    "strict": True,
+                },
+            ),
+            (
+                range(100),
+                {
+                    "type": "static",
+                    "value": list(range(100)),
+                    "display": "typeahead",
+                    "strict": True,
+                },
+            ),
+            (
+                {"value": [1, 2, 3]},
+                {
+                    "type": "static",
+                    "value": [1, 2, 3],
+                    "display": "select",
+                    "strict": True,
+                },
+            ),
+            (
+                {"value": {"a": [1, 2], "b": [3, 4]}, "key_reference": "${y}"},
+                {
+                    "type": "static",
+                    "value": {"a": [1, 2], "b": [3, 4]},
+                    "display": "select",
+                    "strict": True,
+                    "details": {"key_reference": "y"},
+                },
+            ),
+            (
+                "http://myhost:1234",
+                {
+                    "type": "url",
+                    "value": "http://myhost:1234",
+                    "display": "typeahead",
+                    "strict": True,
+                    "details": {"address": "http://myhost:1234", "args": []},
+                },
+            ),
+            (
+                "my_command",
+                {
+                    "type": "command",
+                    "value": "my_command",
+                    "display": "typeahead",
+                    "strict": True,
+                    "details": {"name": "my_command", "args": []},
+                },
+            ),
+            (
+                {"type": "command", "value": {"command": "my_command"}},
+                {
+                    "type": "command",
+                    "value": {"command": "my_command"},
+                    "display": "select",
+                    "strict": True,
+                    "details": {"name": "my_command", "args": []},
+                },
+            ),
+        ],
+    )
+    def test_choices(self, cmd, choices, expected):
+        generated = process_choices(choices)
+
+        assert generated.type == expected["type"]
+        assert generated.value == expected["value"]
+        assert generated.display == expected["display"]
+        assert generated.strict == expected["strict"]
+        assert generated.details == expected.get("details", {})
+
+    @pytest.mark.parametrize(
+        "choices",
+        [
+            # No value
+            {"type": "static", "display": "select"},
+            # Invalid type
+            {"type": "Invalid Type", "value": [1, 2, 3], "display": "select"},
+            # Invalid display
+            {"type": "static", "value": [1, 2, 3], "display": "Invalid display"},
+            # Command value invalid type
+            {"type": "command", "value": [1, 2, 3]},
+            # Static value invalid type
+            {"type": "static", "value": "This should not be a string"},
+            # No key reference
+            {"type": "static", "value": {"a": [1, 2, 3]}},
+            # Parse error
+            {"type": "command", "value": "bad_def(x="},
+            # Just wrong
+            1,
+        ],
+    )
+    def test_choices_error(self, cmd, choices):
+        with pytest.raises(PluginParamError):
+            process_choices(choices)
